+++ app/org/bigraph/bigsim/ctlspec/atom/False.java
@@ -1,7 +1,7 @@
 package org.bigraph.bigsim.ctlspec.atom;
 
-import org.bigraph.bigsim.ctlspec.Formula;
-import org.bigraph.bigsim.ctlspec.FormulaVisitor;
+import visitor.Formula;
+import visitor.FormulaVisitor;
 import org.bigraph.bigsim.ctlspec.operator.Not;
 
 public class False implements Formula {
