+++ app/org/bigraph/bigsim/ctlspec/atom/True.java
@@ -1,7 +1,7 @@
 package org.bigraph.bigsim.ctlspec.atom;
 
-import org.bigraph.bigsim.ctlspec.Formula;
-import org.bigraph.bigsim.ctlspec.FormulaVisitor;
+import visitor.Formula;
+import visitor.FormulaVisitor;
 
 public class True implements Formula {
     public static True True(){
