+++ app/org/bigraph/bigsim/ctlspec/operator/AG.java
@@ -1,7 +1,7 @@
 package org.bigraph.bigsim.ctlspec.operator;
 
-import org.bigraph.bigsim.ctlspec.Formula;
-import org.bigraph.bigsim.ctlspec.FormulaVisitor;
+import visitor.Formula;
+import visitor.FormulaVisitor;
 
 import java.util.Objects;
