+++ app/org/bigraph/bigsim/ltlspec/operator/LTLOperatorAnd.java
@@ -1,7 +1,7 @@
 package org.bigraph.bigsim.ltlspec.operator;
 
-import org.bigraph.bigsim.ltlspec.LTLFormula;
-import org.bigraph.bigsim.ltlspec.LTLFormulaVisitor;
+import visitor.LTLFormula;
+import visitor.LTLFormulaVisitor;
 
 import java.util.Objects;
