+++ app/org/bigraph/bigsim/modelchecker/CTLModelCheckerOnTheFly.java
@@ -3,7 +3,7 @@ package org.bigraph.bigsim.modelchecker;
 import org.bigraph.bigsim.ctlimpl.CheckAUResult;
 import org.bigraph.bigsim.ctlimpl.CheckEUResult;
 import org.bigraph.bigsim.ctlimpl.TarjansDepthFirstSearchData;
-import org.bigraph.bigsim.ctlspec.Formula;
+import visitor.Formula;
 import org.bigraph.bigsim.ctlspec.atom.Atom;
 import org.bigraph.bigsim.ctlspec.atom.False;
 import org.bigraph.bigsim.ctlspec.atom.True;
